Contains the final versions of all GSCID/BRC Core Metadata Standards (v1.3)
The ".xlsx" file contains multiple sheets. Each sheet contains the
Field Name and the associated attributes on separate rows
The ".docx" files cumulatively contain the same information as the
spreadsheet. In this case, the Field Name and Attributes defined for
each piece of the standard have been separated into distinct files.